+++ packages/editor/src/extensions/table/TableActionMenuPlugin.tsx
@@ -210,7 +210,7 @@ const TableActionMenu: React.FC<TableCellActionMenuProps> = ({
 
         const tableSelection = getTableObserverFromTableElement(tableElement);
         if (tableSelection !== null) {
-          tableSelection.clearHighlight();
+          tableSelection.$clearHighlight();
         }
 
         tableNode.markDirty();